Project6-Googlecheck/requirements.txt
cryptography>=3.4.8
argon2-cffi>=21.3.0
requests>=2.25.1
flask>=2.0.0
pytest>=6.0.0 